requirements.txt
astroid==2.2.5
asyncio-nats-client==0.8.2
asyncio-nats-streaming==0.2.0
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.2
opencv-python==4.0.0.21
protobuf==3.7.1
pylint==2.3.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
